% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_binaries_from_set.R
\name{add_binaries_from_set}
\alias{add_binaries_from_set}
\title{Add as many binaries than input columns, taking the value 1 if value in each column is inside a set a values}
\usage{
add_binaries_from_set(
  df,
  vars,
  vals_0,
  vals_1,
  vals_na = NULL,
  default_val = NULL,
  suffix = NULL,
  sep = "."
)
}
\arguments{
\item{df}{A df.table or df.frame: The df to be processed.}

\item{suffix}{A character vector: The suffix to be added to the added binary vars. If NULL, vars are replaced by binary vars.}

\item{sep}{A character vector: separator used to add name suffix to original column names}

\item{cols_character}{A character vector: The name of the columns that will be used to create binary columns.}

\item{value_1}{A character vector: Set of values for which binary columns will be populated with 1 if \code{cols_character} is inside this set.}

\item{value_0}{A character vector: Set of values for which binary columns will be populated with 0 if \code{cols_character} is inside this set.}

\item{value_na}{A character vector: Set of values for which binary columns will be populated with NA if \code{cols_character} is inside this set.}

\item{value_default}{A character vector: Set of values for which binary columns will be populated with otherwise. Will be overriden with 0 if value_0 = NULL and with NA if value_na = NULL}
}
\value{
A df.table or df.frame with the binary columns added.
}
\description{
Useful to create a set of dummy variables from a set of categorical variables if they are in a defined set of choices
For HESPER package, can be used to create binaries for all HESPER items select one columns, to either record if it's a serious problem, undefined, serious problem with subset
Additionnal argument can be used to add a suffix to the column names of the binary created
}
