% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_methods.R
\name{add_hesper_column}
\alias{add_hesper_column}
\title{hespeR S7 Methods
Add a new column to a HesperDefault or HesperEnhanced object}
\usage{
add_hesper_column(x, ...)
}
\description{
hespeR S7 Methods
Add a new column to a HesperDefault or HesperEnhanced object
}
