% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hesper_main.R
\name{add_hesper_main}
\alias{add_hesper_main}
\title{function that compute for each respondent the number of selected items, number of applicable items and collapse the three ordered priorities into one select multiple column}
\usage{
add_hesper_main(
  df,
  col_items = c("hesper_drinking_water", "hesper_food", "hesper_shelter",
    "hesper_toilet", "hesper_clean", "hesper_clean_women", "hesper_clothes_etc",
    "hesper_income_livelihood", "hesper_health", "hesper_health_care_men",
    "hesper_health_care_women", "hesper_distress", "hesper_safety", "hesper_education",
    "hesper_care", "hesper_support", "hesper_separation", "hesper_displaced",
    "hesper_information", "hesper_aid", "hesper_respect", "hesper_movement",
    "hesper_time", "hesper_law", "hesper_gbv", "hesper_drug", "hesper_mental_health", 
  
      "hesper_care_community", "hesper_other"),
  choice_serious = "Serious_problem",
  choice_no_serious = "No_serious_problem",
  choice_dnk = "Do_not_know",
  choice_pnta = "Decline_to_answer",
  choice_na = "Not_applicable_to_household",
  cols_priority = c("hesper_priority_first", "hesper_priority_second",
    "hesper_priority_third"),
  col_name_hesper_top_three = "hesper_top_three_priorities",
  col_gender = "gender",
  choice_male = "men",
  choice_female = "women",
  hesper_item_male = NULL,
  hesper_item_female = c("hesper_clean_women"),
  col_displacement = "pop_group",
  choices_displaced = c("refugees", "idp"),
  choices_non_displaced = c("hosts"),
  hesper_item_displaced = c("hesper_displaced"),
  hesper_item_non_displaced = NULL,
  add_binaries = T,
  add_binaries_subset = T,
  subset = T,
  add_binaries_undefined = T
)
}
\arguments{
\item{df}{dataframe containing the cleaned data}

\item{col_items}{vector of column names containing the hesper items.}

\item{choice_serious}{character string representing the choice for serious problem}

\item{choice_no_serious}{character string representing the choice for no serious problem}

\item{choice_dnk}{character string representing the choice for do not know}

\item{choice_pnta}{character string representing the choice for decline to answer}

\item{choice_na}{character string representing the choice for not applicable to household}

\item{cols_priority}{vector of three column names containing the first, second and third priority hesper items}

\item{col_gender}{character string representing the column name containing the respondent's gender}

\item{choice_male}{character string representing the choice corresponding to male respondent gender}

\item{choice_female}{character string representing the choice corresponding to female respondent gender}

\item{hesper_item_male}{vector of two choice names corresponding to the hesper items specific for male respondent}

\item{col_displacement}{character string representing the column name containing the respondent's displacement status}

\item{hesper_item_displaced}{vector of two choices name coresponding to the hesper items specific for displaced respondent}

\item{add_binaries}{logical value indicating whether to add binary variables for each hesper item}

\item{add_binaries_subset}{logical value indicating whether to add binary variables for each hesper item that are applicable to only a subset of the population}

\item{subset}{logical value indicating whether to add binary subset variables for all top three priority child columns that are applicable to only a subset of the population}

\item{hesper_item_femal}{vector of two choice names corresponding to the hesper items specific for female respodent}

\item{choice_displaced}{character string representing the choice corresponding to displaced respondent}

\item{choice_not_displaced}{character string representing the choice corresponding to not displaced respondent}

\item{col_hesper_subset}{character string representing the column name containing the respondent's hesper items that are applicable only to a gender/displacement status subset. should match column names in the dataset}

\item{hesper_item_not_displaced}{vector of two choices name coresponding to the hesper items specific for not displaced respondent}
}
\value{
dataframe with the number of selected items, the number of applicable items and the three ordered priorities collapsed into one select multiple column
}
\description{
The function computes the number of selected items for each respondent, the number of applicable items and collapse the three ordered priorities into one select multiple column
}
