% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hesper_top3.R
\name{add_hesper_top3}
\alias{add_hesper_top3}
\title{function that compute for each respondent the number of selected items, number of applicable items and collapse the three ordered priorities into one select multiple column}
\usage{
add_hesper_top3(
  df,
  vars_priority = c("hesper_priority_first", "hesper_priority_second",
    "hesper_priority_third"),
  var_hesper_top_three = "hesper_top_three_priorities",
  sv_l = list(displaced = list(hesper_vars = c("hesper_displaced"), subset_var =
    "pop_group", subset_vals = c("refugees", "idp")), resp_gender_female =
    list(hesper_vars = c("hesper_clean_female"), subset_var = "resp_gender", subset_vals
    = c("female")))
)
}
\arguments{
\item{df}{(``) dataframe containing the cleaned data}

\item{vars_priority}{(``) vector of three column names containing the first, second and third priority hesper items}

\item{var_hesper_top_three}{(``) character string representing the new column name containing the three ordered priorities collapsed into one select multiple column}

\item{sv_l}{(``) list of named lists for each element should be named and containing in hesper_vars the list of hesper item(s) asked only to the subset, in subset_var the name of the column which defines the subset (for example population_group or respondent_gender) and in subset_vals the set of values contained in subset_var defining the relevant subset}

\item{col_items}{(``) vector of column names containing the hesper items.}

\item{choice_serious}{(``) character string representing the choice for serious problem}

\item{choice_no_serious}{(``) character string representing the choice for no serious problem}

\item{choice_dnk}{(``) character string representing the choice for do not know}

\item{choice_pnta}{(``) character string representing the choice for decline to answer}

\item{choice_na}{(``) character string representing the choice for not applicable to household}
}
\value{
(``) dataframe with the number of selected items, the number of applicable items and the three ordered priorities collapsed into one select multiple column
}
\description{
The function computes the number of selected items for each respondent, the number of applicable items and collapse the three ordered priorities into one select multiple column
}
