% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_top3.R
\name{add_top3}
\alias{add_top3}
\title{Add a new column with the top three priorities}
\usage{
add_top3(df, new_var, vars_unite)
}
\arguments{
\item{df}{input dataframe}

\item{new_var}{name of the new column}

\item{vars_unite}{vector with column names corresponding to top 1/2/3 priority to combine in unordered top three priorities}
}
\value{
dataframe with new column
}
\description{
Add a new column with the top three priorities
}
