% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.R
\name{apply_hesper_list_cat}
\alias{apply_hesper_list_cat}
\title{Create category-level HESPER vectors from HesperListEnhanced}
\usage{
apply_hesper_list_cat(self)
}
\arguments{
\item{self}{(\code{HesperListEnhanced}) A HesperListEnhanced object.}
}
\value{
(\code{list[character]}) A named list of character vectors representing category-level aggregation (including binaries).
}
\description{
Create category-level HESPER vectors from HesperListEnhanced
}
\details{
This function creates category-level HESPER vectors by aggregating individual
HESPER variables within each category using hierarchical logic:
\itemize{
\item "serious_problem" if any variable is "serious_problem"
\item "pnta" if any variable is "pnta" (and none are "serious_problem")
\item "dnk" if any variable is "dnk" (and none are "serious_problem" or "pnta")
\item "no_serious_problem" if ALL variables are "no_serious_problem"
\item "not_applicable" if ALL variables are "not_applicable"
}
}
\keyword{internal}
