% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.apply_hesper_list_sl.R
\name{apply_hesper_list_sl}
\alias{apply_hesper_list_sl}
\title{Get cleaned hesper_list after applying skip logic}
\usage{
apply_hesper_list_sl(self)
}
\arguments{
\item{self}{A HesperListEnhanced object}
}
\value{
The cleaned hesper_list
}
\description{
Get cleaned hesper_list after applying skip logic
}
