% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_methods.R
\name{apply_skip_logic}
\alias{apply_skip_logic}
\title{Apply skip logic to a HesperDefault or HesperEnhanced object using an SL object}
\usage{
apply_skip_logic(x, sl, ...)
}
\description{
Apply skip logic to a HesperDefault or HesperEnhanced object using an SL object
}
