% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert HesperList to data.frame}
\usage{
as_data_frame(x, bins = TRUE)
}
\arguments{
\item{x}{(\code{HesperListEnhanced}) A HesperListEnhanced object.}

\item{bins}{(\code{logical[1]}) If \code{TRUE}, includes binary columns for each HESPER option in the data.frame.}
}
\value{
(\code{data.frame}) With columns named by \code{hesper_var} and values from \code{hesper_vals} (and optionally binaries). An attribute \code{allow_missing} is added to indicate if missing values are allowed for each variable.
}
\description{
Returns a data.frame representation of a HesperList object.
}
