% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dupes}
\alias{check_dupes}
\title{Check duplicate values in a vector}
\usage{
check_dupes(vec, property = "vector")
}
\arguments{
\item{vec}{(\code{vector}) The vector to check for duplicates.}

\item{property}{(\code{character[1]}) The name of the property being checked (default: 'vector').}
}
\value{
(\code{TRUE | error}) TRUE if no duplicates are found, otherwise throws an error with a message listing the duplicate values.
}
\description{
This function checks for duplicate values in a vector and throws an error if any duplicates are found.
}
\keyword{internal}
