% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_list_class}
\alias{check_list_class}
\title{Check class of a list of items}
\usage{
check_list_class(self, class_name)
}
\arguments{
\item{self}{(\verb{list[1+]}) A list of items to check.}

\item{class_name}{(\code{character[1]}) The name of the class to check against.}
}
\value{
(\code{TRUE | error}) TRUE if all elements are of the specified class, otherwise throws an error.
}
\description{
This function checks if all elements in a list are of a specified class. If any element is not of the specified class, it returns an error message.
}
\keyword{internal}
