% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_sv_l}
\alias{check_sv_l}
\title{Check if lists of subset are in the right format}
\usage{
check_sv_l(sv_l, df, hesper_vars)
}
\arguments{
\item{sv_l}{A list of subset dictionaries}
}
\value{
A stop statement
}
\description{
Check if lists of subset are in the right format
}
