% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_vars_class_in_df}
\alias{check_vars_class_in_df}
\title{Check if variables are of any given class(es)}
\usage{
check_vars_class_in_df(df, vars, class)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of variable names}

\item{class}{A vector of classes}
}
\value{
A stop statement
}
\description{
Check if variables are of any given class(es)
}
