% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_vecs_in_set}
\alias{check_vecs_in_set}
\title{Check if items of vectors in a list are a set of allowed values}
\usage{
check_vecs_in_set(l_x, set, allow_missing = FALSE)
}
\arguments{
\item{l_x}{(\verb{list_named_vector[1+]}) A list of named vectors to check.}

\item{set}{(\verb{vector[1+]}) Allowed values that the variables should contain.}

\item{allow_missing}{(\code{logical[1]}) Whether to allow missing values (default: FALSE).}
}
\value{
(\code{TRUE | error}) TRUE if all values in the specified variables are in the set, otherwise throws an error.
}
\description{
This function checks if the values in specified variables of a data frame are all within a given set of allowed values. If any values are not in the set, it generates an error message detailing the invalid values in which vars and the expected set.
}
\keyword{internal}
