% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_priorities_subset.R
\name{clean_top_priorities_subset}
\alias{clean_top_priorities_subset}
\title{Clean top one/two/three priorities child binary columns with NA for subset of data}
\usage{
clean_top_priorities_subset(
  data,
  col_prio = c("hesper_priority_first", "hesper_priority_second",
    "hesper_priority_third"),
  subset_cols_vals = list(pop_group = list(displaced = list(subset_val = c("idp",
    "refugees"), col_val = c("hesper_displaced"))), resp_gender = list(female =
    list(subset_val = c("female"), col_val = c("hesper_clean_women")))),
  sep.val = "."
)
}
\arguments{
\item{data}{A data frame or data table}

\item{col_prio}{A character vector of length 3 with the names of the top 1/2/3 priority columns}

\item{subset_cols_vals}{A list of list with the names of the columns to subset the data and the values to subset the data
This list be a nested list with names of the subset columns as first layer, with as many names as subset dimension to clean for
For each subset dimension, another list with as many values than subset groups that needs to be cleaned with
then two list elements linking to all values corresponding to the subset in data set followed by the choice value for the parent target columns that needs to be cleaned with NA for the corresponding subset}
}
\value{
A data frame or data table with the top 1/2/3 priority columns replaced with \code{NA} for the specified subset of the data
}
\description{
This function replaces the top 1/2/3 binary columns (which correspond to choices) with \code{NA} for a specified subset of the data.
}
\examples{

## Expand the top first second and third priority columns into child binaries and then clean the top priorities child columns for the subset of data
hesper_dat_clean <- hesper_dat \%>\% 
  ## expand the top priorities columns to have binaries for each choice
  expand_bin(c("hesper_priority_first", "hesper_priority_second", "hesper_priority_third")) \%>\%
  ## clean the top priorities child columns for the subset of data (standard list of subset and choices)
  clean_top_priorities_subset(data = .,
                              col_prio = c("hesper_priority_first", "hesper_priority_second", "hesper_priority_third"))

## check that it did clean the parent columns correctly:
hesper_dat_clean \%>\% 
  filter(pop_group \%in\% c("idp", "refugees")) \%>\% select(pop_group, contains("displaced")) \%>\% distinct

}
