% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_binary_vectors}
\alias{create_binary_vectors}
\title{Create binary vectors for categorical comparisons}
\usage{
create_binary_vectors(vec_chr, ref_chr)
}
\arguments{
\item{vec_chr}{(\verb{character[1+]}) Character vector to compare against reference values}

\item{ref_chr}{(\verb{character[1+]}) Character vector of possible values to create binaries for}
}
\value{
(\code{list}) Named list where each element is a binary integer vector indicating
which positions in vec_chr match each reference value
}
\description{
Create binary vectors for categorical comparisons
}
\keyword{internal}
