% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperList.as_data_frame.R
\name{from_data_frame}
\alias{from_data_frame}
\title{Convert data.frame to HesperList}
\usage{
from_data_frame(df, allow_missing = NULL)
}
\arguments{
\item{df}{(\code{dat}) A data.frame with one column per HESPER variable, each column a character vector of options.}

\item{allow_missing}{Optional logical vector, or will use attr(df, "allow_missing") if present.}
}
\value{
A HesperList object.
}
\description{
Convert data.frame to HesperList
}
