% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hespeR-package.R
\docType{package}
\name{hespeR-package}
\alias{hespeR}
\alias{hespeR-package}
\title{hespeR: R Handle of The Humanitarian Emergency Settings Perceived Needs Scale (HESPER)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gnoblet.github.io/hespeR/}
}

}
\author{
\strong{Maintainer}: Guillaume Noblet \email{gnoblet@zaclys.net}

Authors:
\itemize{
  \item Raphael Bacot \email{raphael.bacot@impact-initiatives.org}
}

}
\keyword{internal}
