% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hesper_dat}
\alias{hesper_dat}
\title{Some Hesper  dummy data}
\format{
A list of variables, aka a data.frame
}
\usage{
hesper_dat
}
\description{
A dataset with hesper columns
}
\keyword{datasets}
