% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hesper_opts}
\alias{hesper_opts}
\title{Allowed HESPER response levels}
\format{
character vector
}
\usage{
hesper_opts
}
\description{
A character vector of all allowed response values for HESPER items.
}
\keyword{internal}
