% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hesper_vars}
\alias{hesper_vars}
\title{All possible HESPER item column names}
\format{
character vector
}
\usage{
hesper_vars
}
\description{
A character vector listing all possible HESPER item columns for use in hespeR data structures and validation.
}
\keyword{internal}
