% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_classes.R
\name{HesperDefault}
\alias{HesperDefault}
\title{S7 class for hespeR data.frames (default)}
\usage{
HesperDefault(
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  hesper_opts = character(0)
)
}
\description{
S7 class for hespeR data.frames (default)
}
\section{Fields}{

\describe{
\item{\code{data}}{The main data as a data.frame.}

\item{\code{hesper_opts}}{Allowed levels for HESPER items (character).}
}}

