% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.R
\name{HesperListEnhanced}
\alias{HesperListEnhanced}
\title{S7 class for hespeR list of vectors with skip logic}
\usage{
HesperListEnhanced(hesper_list = list(), SL = list(), other_list = list())
}
\arguments{
\item{hesper_list}{(\code{list[HesperVector]}) A list of \code{HesperList} items, each representing a HESPER item and its allowed options.}

\item{SL}{(\code{list[hespeR::SL]}) \code{SL} objects representing skip logic rules.}

\item{other_list}{(\code{list[vec]}) Other vectors associated with the HESPER list, typically used for skip logic rules or disaggregation.}
}
\value{
(\code{S7_object}) A S7 object of class \code{HesperListEnhanced}, representing a list of HESPER vectors associated with skip logic rules and additional data.
}
\description{
S7 class for hespeR list of vectors with skip logic
}
