% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperPriorities.R
\name{HesperPriorities}
\alias{HesperPriorities}
\title{S7 class for HESPER priorities}
\usage{
HesperPriorities(
  top1 = character(0),
  top2 = character(0),
  top3 = character(0),
  allow_missing = FALSE
)
}
\arguments{
\item{top1}{(\verb{character[1+]}) Top 1 priority}

\item{top2}{(\verb{character[1+]}) Top 2 priority}

\item{top3}{(\verb{character[1+]}) Top3 priority}

\item{allow_missing}{(\code{logical[1]}) Whether to allow missing values (NA) in the priority vectors. Default is FALSE.}
}
\value{
(\code{S7_object}) A S7 'HesperPriorities' object representing HESPER priorities.
}
\description{
S7 class for HESPER priorities
}
\details{
This class is used to represent and validate HESPER priorities data. At the end of the HESPER module, respondents are asked to rank their top three priorities from the list of items for which they responded they had a serious problem. This class ensures that the provided priorities are valid according to the following rules:
\itemize{
\item Each of \code{top1}, \code{top2}, and \code{top3} must be a character vector of the same length, representing the priorities for each respondent.
\item Each priority must correspond to a valid HESPER item. The function \code{hesper_vars()} provides the list of valid HESPER items.
\item Each priority must be unique within each respondent. That is, no respondent can have the same item ranked in multiple priority positions.
\item The \code{allow_missing} property determines whether missing values (NA) are allowed in the priority vectors. If \code{TRUE}, NA values are permitted; if \code{FALSE}, all priorities must be valid HESPER items.
}

When a proprerty of \link{HesperListEnhanced}, the validator will also ensure that the priorities correspond to items that were indicated as serious problems in the HESPER list.

The \code{top1_bins}, \code{top2_bins}, and \code{top3_bins} properties are getters that generate named lists of binary vectors for each HESPER item, indicating which respondents selected that item as their top 1, top 2, or top 3 priority, respectively.
}
