% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperVector.R
\name{HesperVector}
\alias{HesperVector}
\title{S7 class for hespeR vectors}
\usage{
HesperVector(
  hesper_var = character(0),
  hesper_vals = character(0),
  allow_missing = FALSE
)
}
\arguments{
\item{hesper_var}{(\code{character[1]}) A single HESPER item name.}

\item{hesper_vals}{(\verb{character[1+]}) A vector of allowed options for the HESPER item.}

\item{allow_missing}{(\code{logical[1]}) Whether missing options are allowed for this HESPER item.}
}
\value{
(\code{S7_object}) A S7 \code{HesperVector} object representing a HESPER item and its allowed options.
}
\description{
S7 class for hespeR vectors
}
\details{
The \code{HesperVector} class is designed to encapsulate the core properties of HESPER data. It includes the name of a HESPER item (\code{hesper_var}) and a vector of allowed options for that item (\code{hesper_vals}).
The class validates that the \code{hesper_var} property is a valid Hesper item (see \link{hesper_vars}) and that all values in \code{hesper_vals} are within the allowed options (see specified in \link{hesper_opts}).
}
