% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_not_empty.R
\name{is_not_empty}
\alias{is_not_empty}
\title{Check if a vector is not empty, i.e. not null, positive length and not filled with NA}
\usage{
is_not_empty(x)
}
\arguments{
\item{x}{(\code{vector}) A vector}
}
\value{
(\code{logical}) TRUE if the vector is not empty, FALSE otherwise
}
\description{
Check if a vector is not empty, i.e. not null, positive length and not filled with NA
}
