% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_missing_vars}
\alias{msg_missing_vars}
\title{Standardized error message for missing vars in a data frame}
\usage{
msg_missing_vars(df, vars, property = NULL)
}
\arguments{
\item{df}{(\code{character[1]}) Name of a data frame.}

\item{vars}{(\code{character}) Names of missing variables.}

\item{property}{(\code{character[1] | NULL}) Name of the property being checked (default: NULL).}
}
\value{
(\code{character[3]}) Suitable for use with \code{rlang::abort()}.
}
\description{
Generates a detailed error message for when required variables are missing from a data frame.
}
\seealso{
Other msg: 
\code{\link{msg_invalid_values}()}
}
\concept{msg}
\keyword{internal}
