% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{notify}
\alias{notify}
\title{Custom function that is either warn or stop depending on logical parameter warn}
\usage{
notify(warn = FALSE, ...)
}
\arguments{
\item{warn}{A logical: If TRUE, the message will be a warning. Default is FALSE.}
}
\description{
Custom function that is either warn or stop depending on logical parameter warn
}
