% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hesper_opts.R
\name{rec_hesper_opts}
\alias{rec_hesper_opts}
\title{Recode response options to allowed values for HESPER items}
\usage{
rec_hesper_opts(df, hesper_vars, old_hesper_opts)
}
\arguments{
\item{df}{(\verb{data.frame[,1+]}) A data frame containing HESPER items.}

\item{hesper_vars}{(\verb{character[1+]}) Names of HESPER item column names to recode.}

\item{old_hesper_opts}{(\code{character_named[5]}) Response options mapped to allowed response options for HESPER items. Names are allowed options and values are the old options to recode from.}
}
\value{
(\verb{data.frame[,1+]}) A data frame with HESPER items recoded to the allowed response options.
}
\description{
This function recodes response options in HESPER items to a standardized set of allowed values (see \link{hesper_opts} for the allowed response options).
}
