% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hesper_vars.R
\name{rec_hesper_vars}
\alias{rec_hesper_vars}
\title{Rename HESPER item columns to standardized names}
\usage{
rec_hesper_vars(df, old_hesper_vars)
}
\arguments{
\item{df}{(\verb{data.frame[,1+]}) A data frame containing HESPER items.}

\item{old_hesper_vars}{(\verb{character_named[1+]}) Column names mapped to allowed HESPER item names. Names are new (standardized) names, values are old names to rename from.}
}
\value{
(\verb{data.frame[,1+]}) A data frame with HESPER item columns renamed to the allowed names.
}
\description{
This function renames HESPER item columns in a data frame to a standardized set of allowed names.
}
