% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_subset.R
\name{replace_na_subset}
\alias{replace_na_subset}
\title{Replace binary columns with NA for a subset of the data}
\usage{
replace_na_subset(
  df,
  subset_col,
  subset_value,
  sep = ".",
  col_parent,
  choice_vals
)
}
\arguments{
\item{df}{(``) A data.table or data.frame: The data to be processed.}

\item{subset_col}{(``) A character vector: The name of the column that defines the subset.}

\item{subset_value}{(``) A character vector: The value in \code{subset_col} for which binary columns will be replaced with `NA`.}

\item{sep}{(``) A character vector: The separator used in the child binary column names (default is ".").}

\item{col_parent}{(``) A character vector: The name of the parent column (usually the question) that defines child binary columns.}

\item{choice_vals}{(``) A character vector: The values corresponding to the child binary columns that should be replaced with \code{NA}.}
}
\value{
(``) A data.table or data.frame with the binary columns replaced by \code{NA} for the relevant subset.
}
\description{
This function replaces binary columns (which correspond to choices) with \code{NA} for a specified subset of the data.
It is typically used when expanding top priority columns into child binary columns, ensuring that child columns for
choices that do not apply to a subset are set to \code{NA}.
}
