% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.R
\name{validate_hle_category_list}
\alias{validate_hle_category_list}
\title{Validate category_list in HesperListEnhanced}
\usage{
validate_hle_category_list(category_list, hl_hesper_vars)
}
\arguments{
\item{category_list}{(\code{list[HesperCategories]}) The category_list to validate}

\item{hl_hesper_vars}{(\code{chr}) Character vector of hesper variables from hesper_list}
}
\value{
(\code{NULL}) NULL if valid, otherwise aborts with error
}
\description{
Validate category_list in HesperListEnhanced
}
\keyword{internal}
