% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.R
\name{validate_hle_priority_list}
\alias{validate_hle_priority_list}
\title{Validate priority_list in HesperListEnhanced}
\usage{
validate_hle_priority_list(priority_list, hesper_list, hl_hesper_vars)
}
\arguments{
\item{priority_list}{(\code{list[HesperPriorities]}) The priority_list to validate}

\item{hesper_list}{(\code{list[HesperVector]}) The hesper_list for validation}

\item{hl_hesper_vars}{(\code{chr}) Character vector of hesper variables from hesper_list}
}
\value{
(\code{NULL}) NULL if valid, otherwise aborts with error
}
\description{
Validate priority_list in HesperListEnhanced
}
\keyword{internal}
