% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.R
\name{validate_hle_sl}
\alias{validate_hle_sl}
\title{Validate SL (skip logic) list in HesperListEnhanced}
\usage{
validate_hle_sl(sl_list, hl_hesper_vars, other_list)
}
\arguments{
\item{sl_list}{(\code{list[SL]}) The SL list to validate}

\item{hl_hesper_vars}{(\verb{character[1+]}) Character vector of hesper variables from hesper_list}

\item{other_list}{(\code{list[vec]}) The other_list for subset variable validation}
}
\value{
(\code{NULL}) NULL if valid, otherwise aborts with error
}
\description{
Validate SL (skip logic) list in HesperListEnhanced
}
\keyword{internal}
