% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn.R
\name{warn_var_msg}
\alias{warn_var_msg}
\title{Warn message}
\usage{
warn_var_msg(df, vars, msg)
}
\arguments{
\item{df}{The data frame to check for existing variables}

\item{vars}{A character vector of variable names to check for}

\item{msg}{The message to print}
}
\description{
Warn message
}
\details{
This function checks if any of the variables in \code{vars} are already present in \code{df}.
If any are found, a warning is issued and the variables will be removed from \code{df}.
}
\examples{
df <- data.frame(x = 1, y = 2)
warn_removal(df, c("x", "z"), "Variable(s) will be removed.")
}
