% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn.R
\name{warn_var_no_vals}
\alias{warn_var_no_vals}
\title{Warn if a variable contains no values in a set}
\usage{
warn_var_no_vals(df, var, vals)
}
\arguments{
\item{df}{The data frame to check for existing variables}

\item{var}{A variable name}

\item{set}{A character vector of values to check for}
}
\description{
Warn if a variable contains no values in a set
}
\details{
This function checks if any of the variables in \code{vars} contain no values in \code{set}.
}
