% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_top_three.R
\name{add_top_three}
\alias{add_top_three}
\title{Add a new column with the top three priorities}
\usage{
add_top_three(df, new_col, cols_unite)
}
\arguments{
\item{df}{input dataframe}

\item{new_col}{name of the new column}

\item{cols_unite}{vector with column names corresponding to top 1/2/3 priority to combine in unordered top three priorities}
}
\value{
dataframe with new column
}
\description{
Add a new column with the top three priorities
}
