% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperListEnhanced.apply_hesper_list_sl.R
\name{apply_hesper_list_sl}
\alias{apply_hesper_list_sl}
\title{Apply SL cleaning to HesperListEnhanced object}
\usage{
apply_hesper_list_sl(self)
}
\arguments{
\item{self}{(\code{HesperListEnhanced}) A HesperListEnhanced object}
}
\value{
(\code{HesperListEnhanced}) The cleaned HesperListEnhanced object after applying skip logic
}
\description{
Apply SL cleaning to HesperListEnhanced object
}
