% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsHesperList.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert HesperList to data.frame}
\usage{
as_data_frame(x, ...)
}
\arguments{
\item{x}{(\code{HesperList}) A HesperList object.}
}
\value{
(\code{data.frame}) A data.frame with columns named by \code{hesper_var} and values from \code{hesper_vals}. An attribute \code{allow_missing} is added to indicate if missing values are allowed for each variable.
}
\description{
Returns a data.frame representation of a HesperList object.
}
