% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dupes}
\alias{check_dupes}
\title{Check for duplicate values}
\usage{
check_dupes(vec, msg)
}
\arguments{
\item{vec}{A vector}

\item{msg}{A message}
}
\value{
A stop statement
}
\description{
Check for duplicate values
}
