% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_missing_vars}
\alias{check_missing_vars}
\title{Check for missing variables in a data frame}
\usage{
check_missing_vars(df, vars, property = "hesper_vars")
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of variable names to check}

\item{property}{Name of the property being checked (default: 'hesper_vars').}
}
\value{
A stop statement if any variables are missing, otherwise returns TRUE.
}
\description{
Check for missing variables in a data frame
}
\keyword{internal}
