% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_values_in_set}
\alias{check_values_in_set}
\title{Check if values in a vector are all in a given set of allowed values}
\usage{
check_values_in_set(x, allowed, property = "hesper_opts")
}
\arguments{
\item{x}{A vector of values to check.}

\item{allowed}{A vector of allowed values.}

\item{property}{Name of the property being checked (default: 'hesper_opts').}
}
\value{
TRUE if all values are in the allowed set, otherwise throws an error.
}
\description{
Check if values in a vector are all in a given set of allowed values
}
\keyword{internal}
