% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_vars_in_set}
\alias{check_vars_in_set}
\title{Check if variables in a data frame are in a set of allowed values}
\usage{
check_vars_in_set(df, vars, set)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of column names (quoted)}

\item{set}{A vector of values}
}
\value{
A stop statement if any values are not in the set
}
\description{
This function checks if the values in specified variables of a data frame are all within a given set of allowed values. If any values are not in the set, it generates an error message detailing the invalid values in which vars and the expected set.
}
\keyword{internal}
