% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/internals.R
\name{check_vars_in_set}
\alias{check_vars_in_set}
\title{Check}
\usage{
check_vars_in_set(
  df,
  vars,
  set,
  main_message = "All vars must be in the following set: "
)

check_vars_in_set(
  df,
  vars,
  set,
  main_message = "All vars must be in the following set: "
)

check_vars_in_set(
  df,
  vars,
  set,
  main_message = "All vars must be in the following set: "
)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of column names (quoted)}

\item{set}{A vector of values}

\item{main_message}{A main message}
}
\value{
A stop statement

A stop statement

A stop statement
}
\description{
Check

Stop statement values are not in set

Stop statement values are not in set
}
