% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_vector_class}
\alias{check_vector_class}
\title{Check class of items of a vector}
\usage{
check_vector_class(vec, class_name, property = "vec")
}
\arguments{
\item{vec}{(\verb{vector[1+]}) A vector of items to check.}

\item{class_name}{(\code{character[1]}) The name of the class to check against.}
}
\value{
(\code{TRUE | error}) TRUE if all items are of the specified class, otherwise throws an error.
#' @keywords internal
}
\description{
This function checks if all items in a vector are of a specified class. If any item is not of the specified class, it returns an error message.
}
