% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsHesperList.R
\name{from_data_frame}
\alias{from_data_frame}
\title{Convert data.frame to HesperList}
\usage{
from_data_frame(df, allow_missing = NULL)
}
\arguments{
\item{df}{(\verb{data.frame[, 1+]}) A data.frame with one column per HESPER variable, each column a character vector of options.}

\item{allow_missing}{(\verb{logical[1+]}) Indicates whether missing values are allowed for each variable in the data.frame.}
}
\value{
(\code{HesperList}) A S7 HesperList object.
}
\description{
Convert data.frame to HesperList
}
