% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hesper_opts}
\alias{hesper_opts}
\title{Allowed HESPER response options}
\format{
character vector
}
\usage{
hesper_opts
}
\description{
A character vector of all allowed response options for HESPER items: 'serious_problem', 'no_serious_problem', 'dnk', 'pnta', and 'not_applicable'.
}
\keyword{internal}
