% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hesper_opts}
\alias{hesper_opts}
\title{All allowed response options for HESPER items}
\usage{
hesper_opts(...)
}
\arguments{
\item{...}{(\code{character[0,5]}) Optional options to validate. If none provided, all allowed options are returned.}
}
\value{
(\code{character[5]}) All allowed HESPER response options.
}
\description{
This character vector contains all allowed response options for HESPER items, which can be used for validation or reference in the hespeR package These are: 'serious_problem', 'no_serious_problem', 'dnk', 'pnta', and 'not_applicable'.
}
