% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hesper_vars}
\alias{hesper_vars}
\title{All allowed HESPER item column names}
\usage{
hesper_vars()
}
\arguments{
\item{none}{(\code{empty}) No input parameters.}
}
\value{
(\code{character[29]}) All allowed HESPER item names.
}
\description{
A character vector listing all allowed HESPER item columns for use in hespeR data structures and validation or if needed to be accessed easily.
}
