% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_classes.R
\name{HesperDefault}
\alias{HesperDefault}
\title{S7 class for hespeR data.frames (default)}
\usage{
HesperDefault(
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  hesper_opts = character(0)
)
}
\description{
S7 class for hespeR data.frames (default)
}
\details{
The \code{HesperDefault} class is designed to encapsulate the core properties of HESPER data. It includes a data frame containing HESPER items and a character vector of allowed options for those items.
The class validates that:
\itemize{
\item The \code{hesper_opts} property contains exactly 5 allowed options.
\item The \code{data} property is a data frame with columns that match the allowed HESPER variables.
\item All values in the data frame columns are within the allowed options specified in \link{hesper_opts}.
}
}
\section{Fields}{

\describe{
\item{\code{data}}{The main data as a data.frame.}

\item{\code{hesper_opts}}{Allowed levels for HESPER items (character).}
}}

