% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperDefault.R
\name{HesperDefault}
\alias{HesperDefault}
\title{S7 class for hespeR data.frames (default)}
\usage{
HesperDefault(
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{df}{(\verb{data.frame[,1+]}) A data frame containing HESPER items with HESPER response options.}
}
\description{
S7 class for hespeR data.frames (default)
}
\details{
The \code{HesperDefault} class is designed to encapsulate the core properties of HESPER data. It includes a data frame containing HESPER items and a character vector of allowed options for those items.
The class validates that the \code{df} property is a data frame with columns that match the allowed HESPER variables (see \link{hesper_vars}) and all values in the data frame are within the allowed options (see specified in \link{hesper_opts}).
}
