% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_classes.R
\name{HesperEnhanced}
\alias{HesperEnhanced}
\title{S7 class for hespeR data.frames (enhanced)}
\usage{
HesperEnhanced(
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\description{
Inherits from HesperDefault. May have additional properties or methods.
}
