% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HesperList.R
\name{HesperList}
\alias{HesperList}
\title{S7 class for hespeR list of vectors}
\usage{
HesperList(hesper_list = logical(0))
}
\arguments{
\item{hesper_list}{(\code{list[HesperVector]}) A list of \code{HesperList} items, each representing a HESPER item and its allowed options.}
}
\value{
(\code{S7_object}) A S7 object of class \code{HesperList}, representing a list of HESPER vectors.
}
\description{
S7 class for hespeR list of vectors
}
\details{
The \code{HesperList} class is designed to encapsulate a collection of HESPER vectors. Each vector in the list is an instance of the \code{HesperVector} class, representing a specific HESPER item and its allowed options.

The class validates that each item in the list is a valid \code{HesperVector} and that the list is not empty nor has duplicated HESPER items.
}
