% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_invalid_values}
\alias{msg_invalid_values}
\title{Standardized error message for invalid property values}
\usage{
msg_invalid_values(
  invalid,
  allowed,
  property = "hesper_opts",
  full_message = NULL
)
}
\arguments{
\item{invalid}{(\verb{character[1+]}) Invalid values found.}

\item{allowed}{(\verb{character[1+]}) Allowed values.}

\item{property}{(\code{character[1]}) Name of the property being checked (default: 'hesper_opts').}

\item{full_message}{(\code{character[1] | NULL}) Full message that replace the main message 'Invalid values in \{property\}' if not NULL (default: NULL).}
}
\value{
(\code{character[3]}) Suitable for use with \code{rlang::abort()}.
}
\description{
Generates a detailed error message for invalid values in a property, listing the offending values and the allowed set.
}
\seealso{
Other msg: 
\code{\link{msg_missing_vars}()}
}
\concept{msg}
\keyword{internal}
