% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_missing_vars}
\alias{msg_missing_vars}
\title{Standardized error message for missing vars in a data frame}
\usage{
msg_missing_vars(df, vars, property = "hesper_opts")
}
\arguments{
\item{df}{A string name of data frame.}

\item{vars}{A character vector of missing variable names.}

\item{property}{Name of the property being checked (default: 'hesper_vars').}
}
\value{
A named character vector suitable for use with \code{rlang::abort()}.
}
\description{
Generates a detailed error message for when required variables are missing from a data frame.
}
\seealso{
Other msg: 
\code{\link{msg_invalid_values}()}
}
\concept{msg}
\keyword{internal}
