% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hesper_opts.R
\name{rec_hesper_opts}
\alias{rec_hesper_opts}
\title{Recode response options to allowed values for HESPER items}
\usage{
rec_hesper_opts(df, hesper_vars, old_hesper_opts)
}
\arguments{
\item{df}{A data frame containing HESPER items.}

\item{hesper_vars}{A character vector of HESPER item column names to recode.}

\item{old_hesper_opts}{A named character vector of response options mapped to allowed response options for HESPER items. Names are allowed options and values are the old options to recode from.}
}
\value{
A data frame with HESPER items recoded to the allowed response options.
}
\description{
This function recodes response options in HESPER items to a standardized set of allowed values.
}
