% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesper_classes.R
\name{SL}
\alias{SL}
\title{S7 class for skip logic (SL)}
\usage{
SL(
  hesper_var = character(0),
  subset_var = character(0),
  subset_vals = character(0)
)
}
\description{
S7 class for skip logic (SL)
}
\section{Fields}{

\describe{
\item{\code{hesper_var}}{Name of the HESPER variable (character).}

\item{\code{subset_var}}{Name of the subset variable (character).}

\item{\code{subset_vals}}{Values in the subset variable (character).}
}}

